Move plugin and internal definitions to local package
