+++ cabal.project
@@ -1,5 +1,5 @@
 packages:
-  clash-protocols.cabal
+  ./clash-protocols-base, .
 
 tests: True
 
@@ -14,6 +14,16 @@ source-repository-package
   type: git
   location: https://github.com/cchalmers/circuit-notation.git
   tag: 19b386c4aa3ff690758ae089c7754303f3500cc9
+
+package clash-protocols-base
+  -- Reduces compile times by ~20%
+  ghc-options: +RTS -qn4 -A128M -RTS -j4
+
+  -- Workaround for Haddock/CPP #if issues https://github.com/haskell/haddock/issues/1382
+  haddock-options: --optghc="-optP -P"
+
+  -- Don't pollute docs with large tuple instances
+  haddock-options: --optghc=-DHADDOCK_ONLY
 
 package clash-protocols
   -- Reduces compile times by ~20%
