+++ clash-protocols-base/Setup.hs
@@ -1,0 +1,5 @@
+import Distribution.Simple
+import Prelude
+
+main :: IO ()
+main = defaultMain
