+++ clash-protocols-base/clash-protocols-base.cabal
@@ -1,5 +1,5 @@
 cabal-version:       2.4
-name:                clash-protocols
+name:                clash-protocols-base
 synopsis:            a battery-included library for (dataflow) protocols
 Homepage:            https://gitlab.com/martijnbastiaan/clash-protocols
 version:             0.1
@@ -8,17 +8,6 @@ license:             BSD-2-Clause
 license-file:        LICENSE
 author:              Martijn Bastiaan, QBayLogic B.V.
 maintainer:          Martijn Bastiaan <martijn@hmbastiaan.nl>
-description:
-  Suggested reading order:
-  .
-  * 'Protocols' + README.md
-  * 'Protocols.Df'
-  * 'Protocols.Plugin'
-  * 'Protocols.Hedgehog'
-
-data-files:
-    src/Protocols/Hedgehog.hs
-    src/Protocols/Hedgehog/*.hs
 
 flag ci
     Description: Running on CI? Used to set fail-on-warning flag.
@@ -29,7 +18,7 @@ flag large-tuples
   description:
     Generate instances for classes such as `Units` and `TaggedBundle` for tuples
     up to and including 62 elements - the GHC imposed maximum. Note that this
-    greatly increases compile times for `clash-protocols`.
+    greatly increases compile times for `clash-protocols-base`.
   default: False
   manual: True
 
@@ -91,7 +80,6 @@ common common-options
 
   default-language: Haskell2010
   build-depends:
-    -- GHC >= 9.0
     base >= 4.15.0.0,
     Cabal,
 
@@ -104,7 +92,6 @@ custom-setup
   setup-depends:
     base          >= 4.11 && <5,
     Cabal         >= 2.4,
-    cabal-doctest >= 1.0.1 && <1.1
 
 library
   import: common-options
@@ -114,100 +101,25 @@ library
     CPP-Options: -DLARGE_TUPLES
 
   build-depends:
-    , clash-protocols-base
     , circuit-notation
-    , clash-prelude-hedgehog
     , data-default
     , deepseq
     , extra
     , ghc >= 8.7 && < 9.7
     , hashable
-    , hedgehog >= 1.0.2
-    , lifted-async
-    , mtl
-    , pretty-show
-    , strict-tuple
     , tagged
     , template-haskell
 
-      -- To be removed; we need 'Test.Tasty.Hedgehog.Extra' to fix upstream issues
-    , tasty >= 1.2 && < 1.5
-    , tasty-hedgehog >= 1.2
-
   exposed-modules:
-    Protocols.Avalon.MemMap
-    Protocols.Avalon.Stream
-    Protocols.Axi4.Common
-    Protocols.Axi4.ReadAddress
-    Protocols.Axi4.ReadData
-    Protocols.Axi4.Stream
-    Protocols.Axi4.WriteAddress
-    Protocols.Axi4.WriteData
-    Protocols.Axi4.WriteResponse
-    Protocols.Df
-    Protocols.DfConv
-    Protocols.Hedgehog
-    Protocols.Hedgehog.Internal
-    Protocols.Idle
-    Protocols.Wishbone
-    Protocols.Wishbone.Standard
-    Protocols.Wishbone.Standard.Hedgehog
-    Protocols
-
-    -- 'testProperty' is broken upstream, it reports wrong test names
-    -- TODO: test / upstream ^
-    Test.Tasty.Hedgehog.Extra
-
-  autogen-modules:    Paths_clash_protocols
-
-  other-modules:
-    Paths_clash_protocols
+    Protocols.Cpp
+    Protocols.Internal
+    Protocols.Internal.Classes
+    Protocols.Internal.TaggedBundle
+    Protocols.Internal.TaggedBundle.TH
+    Protocols.Internal.TH
+    Protocols.Internal.Units
+    Protocols.Internal.Units.TH
+    Protocols.Plugin
+    Protocols.Plugin.Internal
 
   default-language: Haskell2010
-
-test-suite unittests
-  import: common-options
-  hs-source-dirs: tests
-  type: exitcode-stdio-1.0
-  ghc-options: -threaded -with-rtsopts=-N
-  main-is: unittests.hs
-  other-modules:
-    Tests.Protocols
-    Tests.Protocols.Df
-    Tests.Protocols.DfConv
-    Tests.Protocols.Avalon
-    Tests.Protocols.Axi4
-    Tests.Protocols.Plugin
-    Tests.Protocols.Wishbone
-
-    Util
-  build-depends:
-    clash-protocols-base,
-    clash-protocols,
-    clash-prelude-hedgehog,
-    data-default,
-    unordered-containers,
-    deepseq,
-    extra,
-    hashable,
-    hedgehog,
-    strict-tuple,
-    tasty >= 1.2 && < 1.5,
-    tasty-hedgehog >= 1.2,
-    tasty-th,
-    tasty-hunit
-
-test-suite doctests
-  import:           common-options
-  type:             exitcode-stdio-1.0
-  default-language: Haskell2010
-  main-is:          doctests.hs
-  hs-source-dirs:   tests
-
-  build-depends:
-    base,
-    clash-protocols-base,
-    clash-protocols,
-    process,
-    doctest
-
