+++ clash-protocols-base/src/Protocols/Internal.hs
