+++ clash-protocols-base/src/Protocols/Internal/Classes.hs
