+++ clash-protocols-base/src/Protocols/Internal/TaggedBundle.hs
