+++ clash-protocols-base/src/Protocols/Internal/Units.hs
