+++ clash-protocols-base/src/Protocols/Plugin.hs
@@ -15,7 +15,9 @@ module Protocols.Plugin (
 import Prelude
 
 -- clash-protocols
-import Protocols
+import Protocols.Internal
+import Protocols.Internal.TaggedBundle
+import Protocols.Internal.Units
 import Protocols.Plugin.Internal
 
 -- circuit-notation
