+++ clash-protocols-base/src/Protocols/Plugin/Internal.hs
